refactor(calculator): printer, reader 인터페이스 및 테스트 코드 수정
