+++ src/main/java/calculator/Application.java
@@ -1,10 +1,13 @@
 package calculator;
 
+import calculator.config.ExecutorConfig;
+import calculator.executor.Executor;
+
 public class Application {
 
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
-        Calculator calculator = new Calculator();
-        calculator.run();
+        Executor executor = new ExecutorConfig().executor();
+        executor.run();
     }
+
 }
