--- src/main/java/calculator/OutputPrinter.java
@@ -1,9 +1,0 @@
-package calculator;
-
-public class OutputPrinter {
-
-    public void println(String message) {
-        System.out.println(message);
-    }
-
-}
