+++ src/main/java/calculator/printer/Printer.java
@@ -1,0 +1,8 @@
+package calculator.printer;
+
+public interface Printer {
+    void print(String message);
+
+    void printWithFormat(String format, int result);
+
+}
