+++ src/main/java/calculator/reader/ConsoleReader.java
@@ -1,9 +1,10 @@
-package calculator;
+package calculator.reader;
 
 import camp.nextstep.edu.missionutils.Console;
 
-public class InputReader {
+public class ConsoleReader implements Reader {
 
+    @Override
     public String read() {
         String command = Console.readLine();
         Console.close();
