+++ src/main/java/calculator/reader/Reader.java
@@ -1,0 +1,6 @@
+package calculator.reader;
+
+public interface Reader {
+    String read();
+
+}
