+++ src/test/java/calculator/ConsolePrinterTest.java
@@ -1,14 +1,16 @@
 package calculator;
 
+import calculator.printer.ConsolePrinter;
+import calculator.printer.Printer;
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
 import java.io.PrintStream;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class OutputPrinterTest {
+class ConsolePrinterTest {
 
-    OutputPrinter printer = new OutputPrinter();
+    Printer printer = new ConsolePrinter();
 
     @Test
     public void println_함수_테스트() {
@@ -20,7 +22,7 @@ class OutputPrinterTest {
         System.setOut(new PrintStream(out));
 
         // when
-        printer.println(message);
+        printer.print(message);
 
         // then
         Assertions.assertThat(out.toString()).isEqualTo(message + newLine);
